chore: move `ContainerPaths` to `crate::types::near::docker_build`
